Rename shared rank filter implementation module
- Rename _percentile_range_filter.py to _rank_filter.py to reflect its shared role across rank-filter APIs.

- Update the internal import and scikit-image license-hook path mappings for the new filename.
