+++ .pre-commit-config.yaml
@@ -152,7 +152,7 @@ repos:
             ^python/cucim/src/cucim/skimage/filters/rank/__init__[.]py$|
             ^python/cucim/src/cucim/skimage/filters/rank/_generic[.]py$|
             ^python/cucim/src/cucim/skimage/filters/rank/_percentile[.]py$|
-            ^python/cucim/src/cucim/skimage/filters/rank/_percentile_range_filter[.]py$|
+            ^python/cucim/src/cucim/skimage/filters/rank/_rank_filter[.]py$|
             ^python/cucim/src/cucim/skimage/filters/rank/tests/test_rank[.]py$|
             ^python/cucim/src/cucim/skimage/filters/ridges[.]py$|
             ^python/cucim/src/cucim/skimage/filters/tests/test_correlate[.]py$|
@@ -442,7 +442,7 @@ repos:
             python/cucim/src/cucim/skimage/filters/rank/__init__[.]py$|
             python/cucim/src/cucim/skimage/filters/rank/_generic[.]py$|
             python/cucim/src/cucim/skimage/filters/rank/_percentile[.]py$|
-            python/cucim/src/cucim/skimage/filters/rank/_percentile_range_filter[.]py$|
+            python/cucim/src/cucim/skimage/filters/rank/_rank_filter[.]py$|
             python/cucim/src/cucim/skimage/filters/rank/tests/test_rank[.]py$|
             python/cucim/src/cucim/skimage/filters/tests/test_correlate[.]py$|
             python/cucim/src/cucim/skimage/filters/tests/test_edges[.]py$|
