+++ python/cucim/src/cucim/skimage/filters/rank/_rank_filter.py
